backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9

PyJWT==2.8.0
python-decouple==3.8
pydantic==2.4.2
typing_extensions==4.8.0